pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markdowncore-ai"
version = "1.0.0"
description = "Markdown CORE AI - Classification, Organisation, Retrieval & Entry for your personal markdown knowledge base"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Piyush Tyagi" }
]
keywords = ["llm", "rag", "knowledge-base", "markdown", "second-brain", "pkm", "obsidian", "ollama", "local-ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]

# Core deps — always installed regardless of backend choice.
# Langchain-ollama included in core: Ollama is the default and recommended backend.
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "pydantic>=2",
    "pyyaml>=6",
    "python-frontmatter>=1.1",
    "markdown-it-py>=3",
    "chromadb>=0.5",
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-ollama>=0.2",
    "langchain-google-genai>=2",
    "unstructured[md]>=0.15",
]

# Optional backend extras — install only what you need.
# pipx install markdowncore-ai            -> Ollama only (default)
# pipx install markdowncore-ai[openai]    -> add OpenAI support
# pipx install markdowncore-ai[all]       -> every backend
[project.optional-dependencies]
openai = [
    "langchain-openai>=0.2",
]
anthropic = [
    "langchain-anthropic>=0.2",
]
gemini = [
    "langchain-google-genai>=2",
]
huggingface = [
    "langchain-huggingface>=0.1",
    "sentence-transformers>=3",
]
all = [
    "langchain-openai>=0.2",
    "langchain-anthropic>=0.2",
    "langchain-google-genai>=2",
    "langchain-huggingface>=0.1",
    "sentence-transformers>=3",
]
dev = [
    "build>=1.0",
    "twine>=5.0",
    "pytest>=8",
    "pytest-cov>=5",
]

[project.urls]
Homepage = "https://github.com/piyush-tyagi-13/mdcore"
Documentation = "https://github.com/piyush-tyagi-13/mdcore/tree/main/docs"
Issues = "https://github.com/piyush-tyagi-13/mdcore/issues"

[project.scripts]
mdcore = "mdcore.cli.commands:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["mdcore*"]

[tool.setuptools.package-data]
"mdcore.docs" = ["*.md"]
